chore: Move toast code
