+++ mobile/src/navigation/screens/settings/ExperimentalSettingsView.tsx
@@ -1,4 +1,4 @@
-import { toast } from "@missingcore/toast";
+import { toast } from "@missingcore/ui/toast";
 import { useNavigation } from "@react-navigation/native";
 import { useTranslation } from "react-i18next";
